+++ CHANGELOG.md
@@ -2,6 +2,45 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [16.1.1](https://github.com/mojaloop/ml-testing-toolkit/compare/v16.1.0...v16.1.1) (2023-03-31)
+
+
+### Bug Fixes
+
+* **mojaloop/#3176:** add short for skipping cb before validation ([#242](https://github.com/mojaloop/ml-testing-toolkit/issues/242)) ([8c8dd99](https://github.com/mojaloop/ml-testing-toolkit/commit/8c8dd9944bac86cbb9c08532a359320ce97f7eba)), closes [mojaloop/#3176](https://github.com/mojaloop/project/issues/3176)
+
+## [16.1.0](https://github.com/mojaloop/ml-testing-toolkit/compare/v16.0.1...v16.1.0) (2023-03-30)
+
+
+### Features
+
+* **mojaloop/#3176:** add no response callback option ([#240](https://github.com/mojaloop/ml-testing-toolkit/issues/240)) ([a709608](https://github.com/mojaloop/ml-testing-toolkit/commit/a70960878fcd00ec6efb5574d9e20e4cf58f4b2f)), closes [mojaloop/#3176](https://github.com/mojaloop/project/issues/3176)
+
+### [16.0.1](https://github.com/mojaloop/ml-testing-toolkit/compare/v16.0.0...v16.0.1) (2023-03-29)
+
+
+### Bug Fixes
+
+* **mojaloop/3256:** mongodb password spl chars ([#239](https://github.com/mojaloop/ml-testing-toolkit/issues/239)) ([c4a56ad](https://github.com/mojaloop/ml-testing-toolkit/commit/c4a56adedb067f7298eb2061bb0d46e0766c2b8b))
+
+## [16.0.0](https://github.com/mojaloop/ml-testing-toolkit/compare/v15.7.0...v16.0.0) (2023-03-01)
+
+
+### ⚠ BREAKING CHANGES
+
+* **mojaloop/#3130:** alter db config and allow password env setting (#237)
+
+### Features
+
+* **mojaloop/#3130:** alter db config and allow password env setting ([#237](https://github.com/mojaloop/ml-testing-toolkit/issues/237)) ([578c714](https://github.com/mojaloop/ml-testing-toolkit/commit/578c7146b837ef4566a75401775c16d1e8a55221)), closes [mojaloop/#3130](https://github.com/mojaloop/project/issues/3130)
+
+## [15.7.0](https://github.com/mojaloop/ml-testing-toolkit/compare/v15.6.3...v15.7.0) (2023-02-03)
+
+
+### Features
+
+* **mojaloop/#3075:** add config option to configure socket io ([#232](https://github.com/mojaloop/ml-testing-toolkit/issues/232)) ([8252111](https://github.com/mojaloop/ml-testing-toolkit/commit/8252111ae483d189be34e21859937f37466f0124)), closes [mojaloop/#3075](https://github.com/mojaloop/project/issues/3075)
+
 ### [15.6.3](https://github.com/mojaloop/ml-testing-toolkit/compare/v15.6.2...v15.6.3) (2023-01-10)